Moved ConvertToVsg class into it's own header/cpp
